scripts/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
pillow>=8.3.0
pytesseract>=0.3.8
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
boto3>=1.18.0
sagemaker>=2.100.0
